backend/requirements.txt
fastapi==0.104.1
python-multipart==0.0.6
uvicorn==0.24.0
boto3==1.29.0
python-dotenv==1.0.0
pydantic==2.4.2
httpx==0.25.0